pyproject.toml
[project]
name = "diary"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aerich>=0.9.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "tortoise-orm[asyncpg]>=0.25.4",
    "uvicorn>=0.41.0",
    "requests>=2.32.5",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "deep-translator>=1.11.4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "python-multipart>=0.0.22",
]

[dependency-groups]
dev = [
    "black>=26.3.0",
    "bs4>=0.0.2",
    "coverage>=7.13.4",
    "deep-translator>=1.11.4",
    "lxml>=6.0.2",
    "mypy>=1.19.1",
    "requests>=2.32.5",
    "ruff>=0.15.5",
]

[tool.aerich]
tortoise_orm = "app.db.database.TORTOISE_CONFIG"
location = "./migrations"
src_folder = "./."
